make things more clean and tidy
