+++ historyClass.py
@@ -64,7 +64,6 @@ class TrackedObject():
     isMoving: bool = field(init=False)
     time_since_update : int = field(init=False)
     max_age : int
-    
 
     def __init__(self, id, first, max_age=30):
         self.objID = id
@@ -86,3 +85,7 @@ class TrackedObject():
             self.time_since_update = 0 
         else:
             self.time_since_update += 1
+        if self.history:
+            # calculating euclidean distance of the first stored detection and last stored detection
+            # this is still hard coded, so its a bit hacky, gotta find a good metric to tell if an object is moving or not
+            self.isMoving = ((self.history[0].X-self.history[-1].X)**2 + (self.history[0].Y-self.history[-1].Y)**2)**(1/2) > 5.0
